+++ packages/falcon-ecommerce-uikit/src/Forms/FormInput.tsx
@@ -1,7 +1,8 @@
 import React from 'react';
 import { Field, getIn, FieldProps } from 'formik';
 import { Box, Label, Input, DefaultThemeProps, ThemedComponentProps, extractThemableProps } from '@deity/falcon-ui';
-import { toGridTemplate } from './helpers';
+import { toGridTemplate } from '../helpers';
+import { Validator, passwordValidator, emailValidator, requiredValidator } from './validators';
 
 export enum FormInputAreas {
   label = 'label',
@@ -38,40 +39,12 @@ const errorLayout: DefaultThemeProps = {
   }
 };
 
-export type Validator = (value: string, label: string) => string | undefined;
-
 export type FormInputProps = {
   label: string;
   name: string;
   validators?: Validator[];
 } & ThemedComponentProps &
   React.InputHTMLAttributes<HTMLInputElement>;
-
-// TODO: move validators to separate file?
-// TODO: when new i18n support is ready use it to translate validation messages
-const validEmailRegex = /^(([^<>()[\]\\.,;:\s@"]+(\.[^<>()[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
-const requiredValidator: Validator = (value, label) => (!value ? `${label} is required` : undefined);
-
-const emailValidator: Validator = value => {
-  if (!value || !validEmailRegex.test(value.toLowerCase())) {
-    return ' Email is invalid';
-  }
-
-  return undefined;
-};
-
-const validPasswordRegex = /^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)/;
-
-const passwordValidator: Validator = value => {
-  if (!value || value.length < 8) {
-    return 'Please enter a password with at least 8 characters';
-  }
-
-  if (!validPasswordRegex.test(value)) {
-    return 'Please use at least one lower, upper case char and digit';
-  }
-  return undefined;
-};
 
 const validateSequentially = (validators: Validator[] = [], label: string) => (value: string) => {
   const firstInvalid = validators.find(validator => validator(value, label) !== undefined);
