+++ packages/falcon-ecommerce-uikit/src/index.ts
@@ -14,3 +14,4 @@ export * from './Blog';
 export * from './Query';
 export * from './Locale';
 export * from './Sidebar';
+export * from './Forms';
